move legacy test to hogql queries
